NeuralNetworks: Implemented Neural Network Manager
Implemented Neural Network Manager class which trains/tests the
neural netowrk with given data and allows to store/load trained
model to file. As part of this change Neural Network class's
init function was modified to make all args optional.

Mnist Classifier is implemented as a subclass to Neural netowrk
Manager which implements the train/test methods to suit the
mnist dataset